UPBGE: Remove m_isShadowDone in KX_Scene.
This variable was planned originally to reduce shadow update.
In the code the images render are done before the screen render, but the
shadow is set to not done before the screen render. So this variable is
only for multiple images render then as the screen render will always
regenerate shadows.

This will works only if we now that there's no modifications between two
images render, but is not the case as it's the user from python with
function "image.refresh()" which call the image render. We don't know
if a user moved an object between two call to "refresh()" and in this
case the m_isShadowDone will introduce a bug by not generating shadows
properly in the second image render.