UPBGE: Fix VAO used with VBO and geometry instancing.
Previously if the display lists was disabled in scene and that the the material
rendered was using geometry instancing and that the scene render storage was
VBO. Then VAO was used for geometry instancing.

To fix it the "else if" after the display list enabled check is replaced by a
single "if".