UPBGE: Fix environment ligthing.
Since the merge the UI was broke, some values of the world was moved in the
environment lighting panel. Also the world values was using GPU_dynamic_uniform
which is replaced by GPU_select_uniform. And finally the zenith and horizon
color was updated only in World and not GPUWorld previously. Then some calls to
GPU_horizon_update_color and GPU_zenith_update_colorare aded in KX_WorldInfo::UpdateBackGround.