UPBGE: Fix python segfault on blenderplayer error.
Previously if the blenderplayer file can't be loaded or any other error was raised.
Then the globalDict which was allocated before was tryed deleted at the end of the
blenderplayer main. But if a error is raised the launcher is not created then python
is not initialized and the globalDict variable can't be deleted.
To fix this issue the globalDict is set to NULL as default and allocated just after
the creation of the launcher to make sure that python was initialized.
We do the same by checking at the end of main if globalDict is non NULL before try
delete it.

This problem is encoutered in blenderplayer only as the embedded player reuse the
existing python context of blender.