UPBGE: Fix memory leak for reinstanced physics shape.
Previously if a triangle shape physics shape was reinstanced it will
always create a new btTriangleIndexVertexArray for the vertex and index
array. But in the case of static physics shapes we didn't check if a
previous btTriangleIndexVertexArray was already existing and need to be
freed. This created a possible huge memory leak.

To fix this issue we just check if m_triangleIndexVertexArray is non NULL
and then free it before allocate a new btTriangleIndexVertexArray in
case of static physics shapes.