UPBGE: Fix KX_CollisionContactPointList memory leak.
Previously if an object which doesn't used collision and had RunCollisionCallbacks
called then the KX_CollisionContactPointList instance passed was not freed.
This case happen when one object of the pair of object in collision doesn't use
collision callbacks.

Here too a big memory leak was possible as this unfreed instance is created for
each collision point each frame.