UPBGE: Avoid use GPU_DYNAMIC_LAMP_SPOTSCALE constantly.
This uniform category was previously able to be used in shader as constants
values. This was too dangerous as the spot scale depends directly on the
lamp object scale.

Because this uniform category is now always dynamic it was renamed :
GPU_DYNAMIC_LAMP_DYNSPOTSCALE to match with other always dynamics values as
GPU_DYNAMIC_LAMP_DYNPERSMAT.