UPBGE: Fix opengl state with text render.
Previously as the text objects were possibly render between two mesh objects
the settings of the text objects was not properly undo.
The first setting not undo is for alpha blend but the issue is more deeper
and come from BLF. The only way found to fix it was to force alpha blend in
DisableForText and undo it a the end of RenderText3D.
The second setting not undo is the lighting and material capability. This
issue is fixed by calling ProcessLighting with uselights to false like a
regular mesh.