UPBGE: Use GPUShader in RAS_Shader.
Previously RAS_Shader was calling all the opengl function needed to create a
shader and manage uniforms/attributs. This code was a duplicated of GPUShader.
Like the other part of the rasterizer we should use as much as possible the
GPUxxxx features from the viewport.

Some function was added to GPUShader to implement it properly in RAS_Shader,
a function to bind a simple float uniform was added, a function to bind an
attribute location and a function to get validation log.

The most complicated issue was for the line number in error messages. Indeed
the GPUShader add some macro at the begin of the header of the shader script
to avoid a gap in line number we use the directive #line 0 just after the
macro if the flag GPU_SHADER_FLAGS_SPECIAL_RESET_LINE is passed to
GPU_shader_create_ex.