UPBGE: Use GPUBuffer in RAS_InstancingBuffer.
Using GPUBuffer allow to reuse existing and frequently maintained code
from the viewport. The only issue encountered was for resize the VBO.
The GPUBuffer is not able to be resized, but reallocated instead.
The reallocation is not a problems as the viewport use a pool of all
freed GPUbuffers and reuse it in some case.
The reallocation is done in RAS_InstancingBuffer::Realloc(). None
performance decrease was noticed.