UPBGE: Fix mipmapping on cube maps.
As the world  background cube map was not using mipmapping then the image used
don't generated mipmap levels. This is not a problem if this same texture is not
reused, but the user will certainly reuse this same texture and in this case
mipmapping will not working in regular materials.

To fix this issue we use mipmaps for background, this has not effect on render.