UPBGE: Create Device/ directory.
This Device directory contains all files from the previous Ghost/
directory which is now deleted. It also contains files from
GameLogic/Joystick.

All sources in this directory are now prefixed of DEV. e.g
DEV_EventConsumer, DEV_Joystick, DEV_InputDevice.

The goal of this directory is to contains everything related to
keyboard/mouse/joystick which are simply a interface from the used
library to a C++ class.