UPBGE: Fix shared reloaded gpu texture in BL_Texture.
Previously the texture reload was based on the fact that only two identities will
own the texture : the material and the BL_Texture. This works perfectly when the
the texture is not reused in the material by a different MTex or reused in an
other material.
In these two last case the reference count of the gpu texture will be always
greater than 1 even if the texture is not used by any materials. This will cause
that the texture is never reloaded.

To fix this issue we don't use the gpu texture reference count in the condition to
reload the texture. Instead we compare the gpu texture (m_gpuTex) in BL_Texture and
the coresponding one to the target in Image. If they are not equal then we call
GPU_texture_from_blender. This solution was though a long time ago but only by
reusing the gpu texture in the Image and not call the conversion function.

It's not because we don't use the reference count to reload that we don't need to
incref or decref it. Doing this decref and incref will make sure that the m_gpuTex
pointer go to a valid gpu texture, maybe unused but valid.