UPBGE: Add a vibration actuator for joystick
This is a simple version of vibration actuator where you can set:

- joystick index (indicates which joystick will rumble)
- strength (the force of the vibration)
- duration (the duration of the vibration)

For the moment we can't set vibration "type" but it is ready to use.

The python API is included (not tested)

For some types of joystick devices (mine for example (xbox controller)),
SDL 2.0.4 is needed. The current version is SDL 2.0.3.