UPBGE: Compile shader when creating 2d filter.
Before we deleted the compilation of the shader because we recreated textures
in the same time. This last operation needed to acces to the current canvas
to get the screen bounds, but this canvas was accesible only at the render time.

At the end the compilation was delayed and caused that if the user want to get
uniforms location an assertion was raised from RAS_Shader::GetUniformLocation.
This because the shader was using an non existing GPUShader.

To solve this issue we just move the compilation in the constructor and let the
texture creation delayed in the first render.