UPBGE: Apply modifiers at game start and restore it at game exit
This saves the mesh as it is before applying modifiers in
KX_BlenderSceneConverter constructor. At ge exit we restore the saved mesh
and delete the mesh created when we apply modifiers.

This is done with a bpy script. I don't know anything about bpy so the
script needs to be improved.

Before using this method I used apply_modifer method and then I used undo
to restore the mesh but it was unstable. This seems to be more stable.

To do: save and restore selected objects when we start the game engine
make a list of modifiers that cause problems in ge runtime and apply only
modifiers that cause problems.
The UI can be changed and designed to be per mesh and not per scene.
Crash in docks level at exit.