UPBGE: Add an API to check if a gameobject is inside frustum
or more exactly if an object is culled or not. Why this wasn't available
before?? Unless it was available and I didn't see it?

This can be very handy in python scripts for example to not update shader
uniforms when the gameobject with shader is culled. Can save a lot of
performances in many cases I guess.