UPBGE: Expose culling state of KX_GameObject in python.
Previously it was impossible to the user to know if an object is culled and
can allow some optimization for the user like for example reduce shader
uniform update.

The user can acces to the culling state through KX_GameObject.culled but, as
the culling is procceed just before the render there's only two place in
python where this culling state is valid. The scene's callbacks:
KX_Scene.pre_draw
KX_Scene.post_draw