Change method to apply modifiers: Per mesh apply
We can't use bpy apply_modifiers method as it suppose to use UNDO to
restore original mesh and UNDO is very unstable and can crash Blender.

So we are forced to use obj.to_mesh method. This means that all modifiers
for the mesh will be applied. We can't apply one modifier for a mesh and
not another one.

So finally I decided to put the UI in mesh tab, in Geometry Data...
Really, I tried many things and I was never satisfied with...

Needs controls to avoid crashes if vertex array length is too
important for example. And needs to avoid a python loop threw 10000
objects if there are 10000 objects in the scene. The objects with
modifiers can be found in c++ I guess and passed to the python script.

This could really simplify things if we can do all in c++...