Fix multisample render with soft particle.
Previously soft particle was not using depth testing to allow using
an offset in the shader. This workaround worked perfectly for non-
multisamples render. Else as the soft particle use a depth texture
which can't be multisample in any way then the test used to make the
soft particle transparent was creating sharping.

The only way to solve this issue is to restore the depth testing and
then remove the depth transparency offset option.