UPBGE: Fix non fixed frame rate time issues.
Previously the full framerate mode incremented the clock time by the
frame time step (= scale / fps) to be sure to proceed a frame with
the algorithm:

frames = int(deltatime * m_ticrate / m_timescale + 1e-6);

But by doing this the clock time and the frame time was totally wrong
and not constant.

To fix this issue in case of non-fixed framerate an other way was used.
First, the clock time is incremented only by the time from the previous
frame as for fixed FPS.
Second, the number of frame to proceed is always set to 1. Third, the
frame time step is set to the increment value of the clock time.

Even if we can use a non-fixed framerate the framerate value is still
used for physics simulation as a time scale. Indeed the function
resolving the fh springs need the framerate value to scale the
velocity applied with the ratio:

currentframestep * framerate

This patch also remove the ambiguity about "fixed time" and "fixed framerate".
The both therms are opposite in the code.

It was tested with:
- physics simulation
- time scaling
- fh spring
- animations
- add object life