UPBGE: Remove viewport bounding box gray clear.
Previously a gray clear was made in the area representing the bounding
box of all camera's viewports of scenes.

This clear was in the most of cases uneeded because the world background
render of the first scene override generally the same zone.

The only case where we saw this gray clear was if the camera's viewports
don't cover the entire area, then the part unused (excluding bars) was
in gray. Also this gray is saw in stereo but as a bug.

This gray clear is removed for two reasons. First it made a bug in stereo
by using a wrong size, and in the best case clearing only one eye and not
the both. Secondly the user wasn't able to control the color used in the
clear.

To remove this clear the function KX_KetsjiEngine::ClearFrame is removed.
And an other clear is made in KX_WorldInfo if there's not a blender world.

Also the framing colorin UI is always showed because is doesn't anymore
concerne only the letter box framing mode.