UPBGE: Explicitly convert stereo mode from blender to game.
Previously the conversion of GameData.stereomode was implicit by just
keeping the same enumeration number and order in blender and in BGE.

But this implicit conversion introduced a bug with the removing of the
stereo dome mode, which break the order of the enumerations and made
impossible to use the 3D TV Top Bottom stereo mode.

To fix this issue the convertion is now done explicitly with a switch.