UPBGE: Fix GLSL inverse support.
The inverse GLSL function is needed for geometry instancing, but this
function is quite recent as it is unsupported before GLSL 1.40.

To avoid forbid geometry instancing for a GLSL function support, an
equivalent of the inverse function is written. This equivalent is put
in gpu_shader_lib.glsl in the marco test : __VERSION__ < 140.

This file is added in both vertex and fragment shader when compiling it.