UPBGE: Implement multiple color layer per vertices.
Previously only one color layer was supported because allowing
8 color layers will increase the memory used per vertices of
7 * 4 bytes = 28 bytes.

But as the new vertices storage system use only the memory needed
based on the UV layers of a mesh, the support of multiple
color layers will reuse the same advantages.

Firts we have to change the conversion to detect the color layers.
To succes this task, MTF_localLayers contains a pointer to the
MCol of the current vertex. This MCol is found in the same time
we look at the UV layers (MTFace).
After this the MTF_localLayers is passed to the function GetRGB
same as an array of 8 int representing the maximum 8 color layers.

But we was founding an attribute index by comparing its name to the
layers names. This can failed when a color layer has the same name
of a uv layer and if the uv and color layer are in the same list
when indices greater to 7 are possible and wrong.

To solve this issue a little struct named RAS_MeshObject::Layer
which is a copy of MTF_localLayers is added. At the contruction of
the mesh all the mesh layers are initialized and passed to the
constructor of the mesh object. Later when the shader are initialized
the mesh layers are get through the function RAS_MeshObject::GetLayers
and passed to the material function GetAttribLayers.

Has we allow to use at maximum 8 color layers, the user should acces
to them, he can do it with KX_VertexProxy.colors which is a list of
color layer containing between 1 and 8 vector.