UPBGE: Clean up python joysticks reinstancing.
Previously the added or removed joystick index was put in an
index array passed to DEV_Joystick::HandleEvents, but this array
was an extra.
It was replaced by iteratin on the second passed array 'addrem'
which is now an array of short instead of bool. Where 0 is nothing,
1 add and 2 remove.

In the same time we let the python joystick initialized in the
first frame thanks to updatePythonJoystick to avoid duplicating
code. The only side effect is that joystick are not accesible before
the first frame in a main python controller script.