UPBGE: Fix keyboard sensor release key delay.
Previously when a key raise a release event any logic brick
handling it was activated the next frame.
This was due to the fact that we was comparing the previous state
with the current state, which mean that the correct value comes
always a frame later.

To fix this issue the comparaison was removed and replaced by a
simple check for events count to enable the sensor. And by a check
with the last state to set the sensor positive or not.
To get the last state the function SCA_InputEvent::End(value)
is introduced.