UPBGE: Fix NULL window when reloading game.
Previously the blenderplayer window was own by the launcher but the
functions instantiating the window was called only ones.

To solve this issue the window management is separated to the launcher.
The LA_PlayerLauncher just had now a take as constructor argument the
main window as it doesn't allocate it.

Finally the main() of the blenderplayer use a pointer to the created
window at the first game run.
Also the the deletion of the window is made just before the deletion
of the system in the blenderplayer main().