UPBGE: Fix animations update and shadows.
Previously since 1d81d517530e53e8545a160078f6a386dc83ead9 (again) the animations
updated during a shaodw render wasn't in the final render.

This was because BL_ActionManager::Update was always called for culled and
non-culled objects. But this same function used a variable named m_prevUpdate
to avoid double update for the same animation time.
The case creating the bug was when an object culled during the shadow render
updates its animation, then the m_prevUpdate is set to the current time and
after in the final render, when trying to update the animation the comparaison
with m_prevUpdate discard the update as exepted for the same object even if
is it non-culled.

To solve this issue we have to know if the action was applied to the object
(similar to if the object is non-culled) and in this case allow discarding
the animation update thanks to m_prevUpdate.

To do it the m_prevUpdate was moved in BL_Action to access to the m_appliedToObject
value in the function BL_Action::Update.