UPBGE: Fix crash when calling Texture_init with material without texture.
Previously if the texture index was pointing to an invalid RAS_Texture
a memory exeption was raised, because the Texture class in function
Texture_init was not checking if the RAS_Texture was valid before acces
to its datas.

To solve this a check is made and a new exeption TextureNotAvail is raised.
This exeption print the error message:
Texture is not available.