UPBGE: Fix render of empty meshes.
Previously only in VA render when a display array was containing none
vertices then the vertex pointer was NULL. This made that the VA
storage RAS_StorageVA, in function TexCoordPtr was trying to acces to
a virtual function  of a NULL instance. This raise a segmentation fault
error.

To avoid this we simply check if the tv pointer is not NULL in TexCoordPtr.

"Why not simply avoid add mesh slot of an empty mesh renderable ?"
This solution was though but rejected for some reasons:
- Empty meshes are an extremely rare case which is more an user error.
  Using a test to now if a mesh slot can be renderable will slown down all
  other mesh slots which are following the rules.
- It's complicated to detect if a display array is empty, it request
  to test it for every render stage or implement function to call when
  finishing the convertion.

It's strongly deprecated to the user to let an empty mesh visible.