UPBGE: Fix crash of material textures without textures.
Previously the texture creation assume that a material
texture (MTex) was always containing a texture (Tex).
But in some rare cases it's wrong.
In the same case, the environment map can be set to NULL.

To fix these issues, we firstly: disable creation of
BL_Texture without a valid texture. Secondly: Check if
the environment map is non NULL before check if the
texture is a cube map, and move this process internally
in BL_Texture constructor.